% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{ejPalette}
\alias{ejPalette}
\title{EJ theme palette}
\usage{
ejPalette(n = 5)
}
\arguments{
\item{n}{the number of colours to output}
}
\description{
This function outputs n colours from the ej palette
}
\seealso{
\code{\link{ejCols}}, \code{\link{valiPalette}}
}
