% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{firaCols}
\alias{firaCols}
\title{Fira theme colours}
\format{
An object of class \code{character} of length 5.
}
\usage{
firaCols
}
\description{
This is a vector with 5 colours to be used in palettes and other visual
elements.
}
\seealso{
\code{\link{firaPalette}}
}
\keyword{datasets}
