% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fira.R
\name{firaPalette}
\alias{firaPalette}
\title{Fira theme palette}
\usage{
firaPalette(n = 5)
}
\arguments{
\item{n}{the number of colours to output}
}
\description{
This function outputs n colours from the fira palette
}
\seealso{
\code{\link{firaCols}}
}
