% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fira.R
\name{firaSave}
\alias{firaSave}
\title{Save plots that use the fira theme}
\usage{
firaSave(filename = "plot.pdf", device = "pdf", ...)
}
\arguments{
\item{filename}{path to a file}

\item{device}{which type of output device to use}

\item{...}{other arguments passed to ggsave}
}
\description{
This function behaves like \code{ggsave} but automatically embeds the fira
font if the output format is pdf.
}
\seealso{
\code{\link{ggplot2::ggsave}}
}
