% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fira.R
\name{setupFont}
\alias{setupFont}
\title{Set up the fira font}
\usage{
setupFont()
}
\description{
This function enables the use of fira in plots and in pdf graphics output. It
is automatically run by theme_fira if necessary
}
\seealso{
\code{\link{theme_fira}}
}
