% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fira.R
\name{theme_fira}
\alias{theme_fira}
\title{Fira theme}
\usage{
theme_fira(family = "Fira Sans")
}
\arguments{
\item{family}{Change the font family. Defaults to Fira Sans.}
}
\value{
ggplot theme
}
\description{
This theme uses Mozilla's Fira Sans as its font. Before using it, run the
\code{setupFont()} function to make it available for both on-screen graphics
and saving to pdf using \code{firaSave()}.
}
\examples{
ggplot(mtcars, aes(x = mpg*0.43, y = wt*0.4535924)) +
  geom_point(size = 2, colour = firaCols[1]) +
  labs(title = "Car weight vs efficiency",
       x = "Efficiency (km/l)",
       y = "Weight (1000 kg)") +
  theme_fira()

}
\seealso{
\code{\link{setupFont}}, \code{\link{firaCols}},
\code{\link{firaPalette}}
}
