% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{valiPalette}
\alias{valiPalette}
\title{Vali theme palette}
\usage{
valiPalette(n = 5)
}
\arguments{
\item{n}{the number of colours to output}
}
\description{
This function outputs n colours from the Vali palette
}
\seealso{
\code{\link{valiCols}}, \code{\link{ejPalette}}
}
