% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fira.R
\name{theme_fira}
\alias{theme_fira}
\title{Fira theme}
\usage{
theme_fira(family = "Fira Sans")
}
\arguments{
\item{family}{Change the font family. Defaults to Fira Sans}
}
\value{
ggplot theme
}
\description{
This theme uses Mozilla's Fira Sans as its font.
Save to pdf using \code{firaSave()}.
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(x = mpg*0.43, y = wt*0.4535924, colour = factor(cyl))) +
  geom_point(size = 2) +
  labs(title = "Car weight vs efficiency",
       subtitle = "Using sensible metrics",
       x = "Efficiency (km/l)",
       y = "Weight (1000 kg)",
       colour = "Cylinders") +
  theme_fira() +
  scale_colour_fira()

}
\seealso{
\code{\link{firaSave}}, \code{\link{scale_fira}}
}
