% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fira.R
\name{theme_fira}
\alias{theme_fira}
\title{Fira theme}
\usage{
theme_fira(family = "Fira Sans", colourPalette = ejPalette)
}
\arguments{
\item{family}{Change the font family. Defaults to Fira Sans.}

\item{colourPalette}{either the function ejPalette or valiPalette
(or any grDevices::colorRampPalette function)}
}
\value{
ggplot theme
}
\description{
This theme uses Mozilla's Fira Sans as its font.
Save to pdf using \code{firaSave()}.
}
\examples{
ggplot(mtcars, aes(x = mpg*0.43, y = wt*0.4535924, colour = cyl)) +
  geom_point(size = 2) +
  labs(title = "Car weight vs efficiency",
       x = "Efficiency (km/l)",
       y = "Weight (1000 kg)") +
  theme_fira()

}
\seealso{
\code{\link{valiPalette}}, \code{\link{ejPalette}}
}
